refactor: resolve eslint warnings
