+++ src/account/components/ApiKeyAdministratorInfo.tsx
@@ -1,7 +1,7 @@
+import { AdministratorRoleName } from "@administration/types";
+import { hasSufficientAdminRole } from "@administration/utils";
 import Alert from "@base/Alert";
 import React from "react";
-import { AdministratorRoleName } from "../../administration/types";
-import { hasSufficientAdminRole } from "../../administration/utils";
 import { useFetchAccount } from "../queries";
 
 /**
