+++ src/hmm/queries.ts
@@ -1,10 +1,10 @@
+import { ErrorResponse } from "@/types/api";
 import {
     keepPreviousData,
     useMutation,
     useQuery,
     useQueryClient,
 } from "@tanstack/react-query";
-import { ErrorResponse } from "../types/types";
 import { fetchHmm, installHmm, listHmms } from "./api";
 import { Hmm, HMMInstalled, HmmSearchResults } from "./types";
