+++ src/indexes/queries.ts
@@ -1,5 +1,5 @@
+import { ErrorResponse } from "@/types/api";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { ErrorResponse } from "../types/types";
 import {
     createIndex,
     findIndexes,
