+++ src/samples/queries.ts
@@ -1,3 +1,4 @@
+import { ErrorResponse } from "@/types/api";
 import {
     keepPreviousData,
     useMutation,
@@ -6,7 +7,6 @@ import {
 } from "@tanstack/react-query";
 import { forEach, map, reject, union } from "lodash-es/lodash";
 import { Label } from "../labels/types";
-import { ErrorResponse } from "../types/types";
 import {
     createSample,
     getSample,
