+++ src/subtraction/queries.ts
@@ -1,10 +1,10 @@
+import { ErrorResponse } from "@/types/api";
 import {
     keepPreviousData,
     useMutation,
     useQuery,
     useQueryClient,
 } from "@tanstack/react-query";
-import { ErrorResponse } from "../types/types";
 import {
     createSubtraction,
     fetchSubtractionShortlist,
