+++ src/types/api.ts
@@ -1,4 +1,14 @@
-/** Contains search results information */
+export type ErrorResponse = {
+    response: {
+        status: number;
+        notFound: boolean;
+        statusText: string;
+        badRequest: boolean;
+        statusCode: number;
+        body: { [key: string]: any };
+    };
+};
+
 export type SearchResult = {
     /** The number of items found */
     found_count: number;
