--- src/types/types.ts
@@ -1,10 +1,0 @@
-export type ErrorResponse = {
-    response: {
-        status: number;
-        notFound: boolean;
-        statusText: string;
-        badRequest: boolean;
-        statusCode: number;
-        body: { [key: string]: any };
-    };
-};
